Move the classes from the com.seudev.util.text package to com.seudev.util.el package
