+++ src/main/java/com/seudev/util/el/ExpressionDefinitions.java
@@ -1,4 +1,4 @@
-package com.seudev.util.text;
+package com.seudev.util.el;
 
 import java.util.SortedMap;
 import java.util.SortedSet;
