--- src/main/java/com/seudev/util/text/package-info.java
@@ -1,4 +1,0 @@
-/**
- * @author Thomás Sousa Silva (ThomasSousa96)
- */
-package com.seudev.util.text;
